AT-13 - Moved ExtendableIterable::assertIterableKeysAndValues() to a trait.
